# Improve performance of math functions