This PR prevents double conversion for math functions.

Benchmark:

```
[MemoryDiagnoser]
public class TestComparision
{
    public object num = "1324729.68374";

    [Benchmark(Baseline = true)]
    public double TestConvertIfNeeded()
    {
        var res = ConvertIfNeeded(num, CultureInfo.InvariantCulture);
        return Convert.ToDouble(res);
    }

    [Benchmark]
    public double TestConvertToDouble()
    {
        return ConvertToDouble(num, CultureInfo.InvariantCulture);
    }

    private static object? ConvertIfNeeded(object? value, MathHelperOptions options)
    {
        return value switch
        {
            char when options is { DecimalAsDefault: true, AllowCharValues: false } => decimal.Parse(value.ToString()!, options.CultureInfo),
            string when options is { DecimalAsDefault: true } => decimal.Parse(value.ToString()!, options.CultureInfo),
            char when options is { AllowCharValues: false } => double.Parse(value.ToString()!, options.CultureInfo),
            string => double.Parse(value.ToString()!, options.CultureInfo),
            bool boolean when options.AllowBooleanCalculation => boolean ? 1 : 0,
            _ => value
        };
    }

    private static double ConvertToDouble(object? value, MathHelperOptions options)
    {
        return value switch
        {
            char => Convert.ToDouble(value.ToString(), options.CultureInfo),
            double => (double)value,
            _ => Convert.ToDouble(value, options.CultureInfo)
        };
    }
}
```

Results when `num` is `double`:
// * Summary *

BenchmarkDotNet v0.14.0, Windows 10 (10.0.19045.5011/22H2/2022Update)
Intel Core i3-6100U CPU 2.30GHz (Skylake), 1 CPU, 4 logical and 2 physical cores
.NET SDK 8.0.206
  [Host]     : .NET 6.0.35 (6.0.3524.45918), X64 RyuJIT AVX2
  DefaultJob : .NET 6.0.35 (6.0.3524.45918), X64 RyuJIT AVX2


| Method              | Mean      | Error     | StdDev    | Median    | Ratio | RatioSD | Allocated | Alloc Ratio |
|-------------------- |----------:|----------:|----------:|----------:|------:|--------:|----------:|------------:|
| TestConvertIfNeeded | 18.791 ns | 0.1535 ns | 0.1436 ns | 18.776 ns |  1.00 |    0.01 |         - |          NA |
| TestConvertToDouble |  5.850 ns | 0.1681 ns | 0.3618 ns |  5.696 ns |  0.31 |    0.02 |         - |          NA |

// * Hints *
Outliers
  TestComparision.TestConvertToDouble: Default -> 4 outliers were removed (9.90 ns..12.64 ns)



Results when `num` is `string`:
// * Summary *

BenchmarkDotNet v0.14.0, Windows 10 (10.0.19045.5011/22H2/2022Update)
Intel Core i3-6100U CPU 2.30GHz (Skylake), 1 CPU, 4 logical and 2 physical cores
.NET SDK 8.0.206
  [Host]     : .NET 6.0.35 (6.0.3524.45918), X64 RyuJIT AVX2
  DefaultJob : .NET 6.0.35 (6.0.3524.45918), X64 RyuJIT AVX2


| Method              | Mean     | Error   | StdDev  | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|-------------------- |---------:|--------:|--------:|------:|--------:|-------:|----------:|------------:|
| TestConvertIfNeeded | 146.5 ns | 2.88 ns | 2.55 ns |  1.00 |    0.02 | 0.0153 |      24 B |        1.00 |
| TestConvertToDouble | 131.6 ns | 2.55 ns | 2.38 ns |  0.90 |    0.02 |      - |         - |        0.00 |

// * Hints *
Outliers
  TestComparision.TestConvertIfNeeded: Default -> 1 outlier  was  removed (168.54 ns)